+++ gss-domain/src/main/java/com/devoops/domain/entity/analysis/AiCharge.java
@@ -1,5 +1,6 @@
 package com.devoops.domain.entity.analysis;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -9,5 +10,5 @@ public class AiCharge {
 
     private final int year;
     private final int month;
-    private final double charge;
+    private final BigDecimal charge;
 }
